Get Data from API

This tool is used retrieved data from the Cal-Adapt API. The tool will add a table to the ArcGIS Pro project that it is run in.

GetDataAPI (in_feature_set, {individual_features}, {catField}, Variable, GCM, Period, historical, rcp45, rcp85, vic, Stat, outTable)

Arguments

in_feature_set
Features to use in your query to the Cal-Adapt API. These features can be a single or series of points, lines, and polygons. The Cal-Adapt API will not allow features to query an area larger that San Bernadino county in California. You can also sketch a feature on the map in ArcGIS Pro and use that graphic as an input to this tool.

individual_features (Optional)
Unchecked all features in the Input feature set will be treated as one feature. Checked all feature will be treated as individual features and each will be used in the query to the API and the results will be combined into a single table with a category field with a unique identifier for each feature.

catField (Optional)
If you choose to process features individually, specify a field in the data set that can be used to identify the unique features in the feature set.

Variable
If you choose to process features individually, specify a field in the data set that can be used to identify the unique features in the feature set.

GCM
Global Climatic Models (GCM) you would like to query using the Cal-Adapt API. A list of GCMs from Cal-Adapt API (https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#global-climate-models-gcm).

Period
Temporal Aggregations you would like to query using the Cal-Adapt API. A list of periods from Cal-Adapt API (https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#period).

historical
Check to include historical data.

rcp45
Check to include RCP 4.5 (Emissions peak around 2040, then decline). To learn more about Representative Concentration Pathways (RCPs) go here.

rcp85
Check to include RCP 8.5 (Emissions continue to rise strongly through 2050 and plateau). To learn more about Representative Concentration Pathways (RCPs) go here.

vic
This is only an option if you choose the 'livneh' GCM.

Stat
Choose what statistic you would like returned from the Cal-Adapt API.

outTable
Specifiy the output path of the table you would like to create with this tool.

Details

This tool is used retrieved data from the Cal-Adapt API. The tool will add a table to the ArcGIS Pro project that it is run in.

See also