LOCA Data

Download the SCRIPPS LOCA Downscaled CMIP5 Climate Projections as NetCDF files from the Cal-Adapt Data Server.

LOCA_Data (Data Type, Climate Model, Climate Scenario, variables, start year, {end year}, outLoc)

Arguments

Data Type
Broad Category

Climate Model
Global Climatic Models (GCM) you would like to query using the CalAdapt API. A list of GCMs from CalAdapt API (https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#global-climate-models-gcm).

Climate Scenario
To learn more about Representative Concentration Pathways (RCPs) go here.

variables
Climatic Variables you would like to download from the Cal-Adapt Data Server.

start year
Choose the first (or only) year to download (Historic: years available 1950-2005, RCP# years available 2006-2100)

end year (Optional)
Choose the last year (if downloading a range) to download (Historic: years available 1950-2006, RCP# years available 2006-2100)

outLoc
Where would you like the data saved.

Details

Download the SCRIPPS LOCA Downscaled CMIP5 Climate Projections as NetCDF files from the Cal-Adapt Data Server.

Daily climate projections for California at a resolution of 1/16° (about 6 km, or 3.7 miles) generated to support climate change impact studies for California’s Fourth Climate Change Assessment. The data, derived from 32 coarse-resolution (~100 km) global climate models from the CMIP5 archive, were bias corrected and downscaled using the Localized Constructed Analogues (LOCA) statistical method. The data cover 1950-2005 for the historical period and 2006-2100 (some models stop in 2099) for two future climate projections. Details are described in Pierce et al., 2018.

See also