Stream Flow Data

Download the SCRIPPS Routed Streamflow Projections as NetCDF files from the Cal-Adapt Data Server.

StreamFlow_Data (Climate Model, Climate Scenario, variables, outLoc)

Arguments

Climate Model
Global Climatic Models (GCM) you would like to query using the Cal-Adapt API. A list of GCMs from Cal-Adapt API (https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#global-climate-models-gcm).

Climate Scenario
To learn more about Representative Concentration Pathways (RCPs) go here.

variables
Climatic Variables you would like to download from the Cal-Adapt Data Server.

outLoc
Where would you like the data saved.

Details

Download the SCRIPPS Routed Streamflow Projections as NetCDF files from the Cal-Adapt Data Server.

Bias-corrected monthly streamflow projections constructed by routing runoff and baseflow outputs from the LOCA VIC runs and bias corrected based on California Department of Water Resources estimates of unimpaired flows. Data is available for 11 locations in California. Details are described in Pierce et al, 2018.

See also